Rename the array implementation to indicate it uses strategies
