+++ src/som/primitives/ast/method_primitives.py
@@ -2,7 +2,7 @@ from som.primitives.primitives import Primitives
 from som.vm.globals import nilObject
 
 from som.vmobjects.abstract_object import AbstractObject
-from som.vmobjects.array           import Array
+from som.vmobjects.array_strategy           import Array
 from som.vmobjects.method          import Method
 from som.vmobjects.primitive       import Primitive
