+++ src/som/primitives/ast/object_primitives.py
@@ -3,9 +3,9 @@ from rpython.rlib.objectmodel import compute_identity_hash
 from som.primitives.primitives import Primitives
 from som.vm.globals import nilObject, falseObject, trueObject
 
-from som.vmobjects.object    import Object  
+from som.vmobjects.object    import Object
 from som.vmobjects.primitive import Primitive
-from som.vmobjects.array     import Array 
+from som.vmobjects.array_strategy     import Array
 
 
 def _equals(ivkbl, rcvr, args):
