+++ src/som/vmobjects/array_strategy.py
